requirements.txt
Flask==2.1.1
Flask-SQLAlchemy==2.5.1
python-dotenv
pytest==6.2.5
pytest-cov==3.0
werkzeug==2.1.1